webserver.rest settings
